Further "agnostify" the main bits: + remove inclusion of rwchc_export.h from rwchcd.h + move hardware config to rwchcd_hardware, with accessors

+++ rwchcd_config.h
@@ -20,7 +20,7 @@ struct s_config * config_new(void);
 void config_del(struct s_config * config);
 int config_init(struct s_config * const config);
 int config_set_building_tau(struct s_config * const config, const time_t tau);
-int config_set_nsensors(struct s_config * const config, const int_fast16_t nsensors);
+int config_set_nsensors(struct s_config * const config, const int_fast8_t nsensors);
 int config_set_tfrostmin(struct s_config * const config, const temp_t tfrostmin);
 int config_set_tsummer(struct s_config * const config, const temp_t tsummer);
 int config_set_outdoor_sensorid(struct s_config * const config, const tempid_t sensorid);
